requirements.txt
python-dotenv
tweepy
firebase-admin